refactor: Split components into separate files
