+++ src/App.js
@@ -1,9 +1,11 @@
 import React, { Component } from 'react';
-import { Route, Link } from 'react-router-dom';
+import { Route } from 'react-router-dom';
 import { debounce } from 'throttle-debounce';
 import * as BooksAPI from './BooksAPI';
 import './App.css';
 // import getAll from './data';
+import ListBooks from './ListBooks';
+import SearchBooks from './SearchBooks';
 
 class BooksApp extends Component {
   bookshelves = [
@@ -90,219 +92,5 @@ class BooksApp extends Component {
     );
   }
 }
-
-class ListBooks extends Component {
-  render() {
-    const { bookshelves, books, onMove } = this.props;
-    return (
-      <div className="list-books">
-        <div className="list-books-title">
-          <h1>MyReads</h1>
-        </div>
-        <Bookcase bookshelves={bookshelves} books={books} onMove={onMove} />
-        <OpenSearchButton />
-      </div>
-    );
-  }
-}
-
-const OpenSearchButton = () => {
-  return (
-    <div className="open-search">
-      <Link to="search">
-        <button>Add a Book</button>
-      </Link>
-    </div>
-  );
-};
-
-const Bookcase = props => {
-  const { bookshelves, books, onMove } = props;
-  return (
-    <div className="list-books-content">
-      <div>
-        {bookshelves.map(shelf => (
-          <Bookshelf
-            key={shelf.key}
-            shelf={shelf}
-            books={books}
-            onMove={onMove}
-          />
-        ))}
-      </div>
-    </div>
-  );
-};
-
-const Bookshelf = props => {
-  const { shelf, books, onMove } = props;
-  const booksOnThisShelf = books.filter(book => book.shelf === shelf.key);
-  // console.log('booksOnThisShelf', booksOnThisShelf);
-  return (
-    <div className="bookshelf">
-      <h2 className="bookshelf-title">{shelf.name}</h2>
-      <div className="bookshelf-books">
-        <ol className="books-grid">
-          {booksOnThisShelf.map(book => (
-            <Book key={book.id} book={book} shelf={shelf.key} onMove={onMove} />
-          ))}
-        </ol>
-      </div>
-    </div>
-  );
-};
-
-const Book = props => {
-  const { book, shelf, onMove } = props;
-  return (
-    <li>
-      <div className="book">
-        <div className="book-top">
-          <div
-            className="book-cover"
-            style={{
-              width: 128,
-              height: 193,
-              backgroundImage: `url(${book.imageLinks &&
-                book.imageLinks.thumbnail})`,
-            }}
-          />
-          <BookshelfChanger book={book} shelf={shelf} onMove={onMove} />
-        </div>
-        <div className="book-title">{book.title}</div>
-        <div className="book-authors">
-          {book.authors && book.authors.join(', ')}
-        </div>
-      </div>
-    </li>
-  );
-};
-
-class BookshelfChanger extends Component {
-  state = {
-    value: this.props.shelf,
-  };
-  handleChange = event => {
-    this.setState({ value: event.target.value });
-    this.props.onMove(this.props.book, event.target.value);
-  };
-  render() {
-    // console.log(this.props.shelf);
-    return (
-      <div className="book-shelf-changer">
-        <select value={this.state.value} onChange={this.handleChange}>
-          <option value="move" disabled>
-            Move to...
-          </option>
-          <option value="currentlyReading">Currently Reading</option>
-          <option value="wantToRead">Want to Read</option>
-          <option value="read">Read</option>
-          <option value="none">None</option>
-        </select>
-      </div>
-    );
-  }
-}
-
-class SearchBooks extends Component {
-  render() {
-    const {
-      searchBooks,
-      myBooks,
-      onSearch,
-      onResetSearch,
-      onMove,
-    } = this.props;
-    // console.log(books);
-    return (
-      <div className="search-books">
-        <SearchBar onSearch={onSearch} onResetSearch={onResetSearch} />
-        <SearchResults
-          searchBooks={searchBooks}
-          myBooks={myBooks}
-          onMove={onMove}
-        />
-      </div>
-    );
-  }
-}
-
-const SearchBar = props => {
-  const { onSearch, onResetSearch } = props;
-  return (
-    <div className="search-books-bar">
-      <CloseSearchButton onResetSearch={onResetSearch} />
-      <SearchBooksInput onSearch={onSearch} />
-    </div>
-  );
-};
-
-const CloseSearchButton = props => {
-  const { onResetSearch } = props;
-  return (
-    <Link to="/">
-      <button className="close-search" onClick={onResetSearch}>
-        Close
-      </button>
-    </Link>
-  );
-};
-
-class SearchBooksInput extends Component {
-  state = {
-    value: '',
-  };
-  handleChange = event => {
-    // this.setState({ value: event.target.value });
-    const val = event.target.value;
-    this.setState({ value: val }, () => {
-      // console.log(val);
-      // if (val.length >= 1) {
-      this.props.onSearch(val);
-      // }
-    });
-  };
-  render() {
-    return (
-      <div className="search-books-input-wrapper">
-        <input
-          type="text"
-          value={this.state.value}
-          placeholder="Search by title or author"
-          onChange={this.handleChange}
-          autoFocus
-        />
-      </div>
-    );
-  }
-}
-
-const SearchResults = props => {
-  const { searchBooks, myBooks, onMove } = props;
-
-  const updatedBooks = searchBooks.map(book => {
-    myBooks.map(b => {
-      if (b.id === book.id) {
-        book.shelf = b.shelf;
-      }
-      return b;
-    });
-    return book;
-  });
-  return (
-    <div className="search-books-results">
-      <ol className="books-grid">
-        {updatedBooks.map(book => (
-          <Book
-            key={book.id}
-            book={book}
-            shelf={book.shelf ? book.shelf : 'none'}
-            onMove={onMove}
-          />
-        ))}
-      </ol>
-    </div>
-  );
-};
 
 export default BooksApp;
