+++ src/SearchBooks.js
@@ -1,0 +1,28 @@
+import React, { Component } from 'react';
+import SearchResults from './SearchResults';
+import SearchBar from './SearchBar';
+
+class SearchBooks extends Component {
+  render() {
+    const {
+      searchBooks,
+      myBooks,
+      onSearch,
+      onResetSearch,
+      onMove,
+    } = this.props;
+    // console.log(books);
+    return (
+      <div className="search-books">
+        <SearchBar onSearch={onSearch} onResetSearch={onResetSearch} />
+        <SearchResults
+          searchBooks={searchBooks}
+          myBooks={myBooks}
+          onMove={onMove}
+        />
+      </div>
+    );
+  }
+}
+
+export default SearchBooks;
